Adjust checkouts of EDK2 repositories
Part of the build process copies the rebuilt ARM Trusted Firmware over
the versions in edk2-non-osi.  This requires cleaning edk2-non-osi before
doing checkout.

Merge checking out of the branches with the rest of the repository
handling.  Unfortunately this adjustment can only handle the case of two
branches, but that is good enough for now.

Be consistent about checking out remote branches.  If "master" is checked
out a local "master" branch will interfere (though this might be good for
some cases).